Downloads/Satmi-Chatbot/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
pydantic==2.11.3
pydantic-settings==2.8.1
langchain==0.3.23
langgraph==0.3.31
httpx==0.28.1
python-dotenv==1.1.0
SQLAlchemy==2.0.39
psycopg[binary]==3.2.6
langgraph-checkpoint-postgres>=2.0.0
pytest==8.3.5
prometheus-client==0.21.1
opentelemetry-api==1.33.1
opentelemetry-sdk==1.33.1
opentelemetry-exporter-otlp-proto-http==1.33.1
redis==5.2.1
